requirements.txt
bentoml==1.1.10
numpy
pandas
scikit-learn
PyJWT
starlette
pydantic
pytest
